src/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
supabase==2.3.4
postgrest==0.13.2
python-dotenv==1.0.0
APScheduler==3.10.4
httpx<0.26,>=0.24
email-validator==2.1.0
